the last update

+++ TimViec/ViewModel/SearchViewModel.cs
@@ -3,17 +3,12 @@
 	public class SearchViewModel
 	{
         public int Id { get; set; }
-
         public string JobName { get; set; }
-
         public string CompanyName { get; set; }
-
-        public decimal Salary { get; set;}
-
+        public int Salary { get; set;}
         public string R1_Language { get; set;}
         public string R2_Language { get; set;}
         public string R3_Language { get; set;}
-
         public string Location { get; set;}
         public string Image { get; set;}
     }
